Fixes for Creator 3.3  - utoa() was added to stdlib in gcc 4.9.3; commented it out.  - build order issues with bootloader resolved  - path order to bootloader resolved (path name now includes 4.9.3)  - added some typecasts to utoa() calls in WString.cpp

+++ Uno_blocks.cydsn/core/WString.cpp
@@ -71,7 +71,7 @@ String::String(unsigned char value, unsigned char base)
 {
 	init();
 	char buf[1 + 8 * sizeof(unsigned char)];
-	utoa(value, buf, base);
+	utoa((long unsigned int)value, buf, (int)base);
 	*this = buf;
 }
